Move test class under producer folder
